size_t strlcat strlcpy strncmp strnstr

+++ ft_strlcpy.c
@@ -6,14 +6,14 @@
 /*   By: raqroca- <raqroca-@student.42.madrid.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/13 13:01:17 by raqroca-          #+#    #+#             */
-/*   Updated: 2026/01/13 13:01:52 by raqroca-         ###   ########.fr       */
+/*   Updated: 2026/01/14 12:04:26 by raqroca-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned	int	ft_strlcpy(char *dest, char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest, char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	len;
+	size_t	i;
+	size_t	len;
 
 	i = 0;
 	len = 0;
